Yang-Mills implemented for group Z/nZ
To see this update at work, run  src\ambient_space\z2_lattice.jl (and not src\ambient_space\z2_visuals.jl). You can select the desired group at the end of the script.

There is now support for 2D Yang-Mills over the groups Z/nZ, for any natural n.

Initial angles are now set to 0.0 (used to be pi/2), as well as the initial energy of the faces (which used to be -beta). This desition is so that the initial conditions work across all groups (it makes no sense to have initial anlge pi/2 forn Z/2Z, since it only supports 0 and pi as angles. The energy of the faces suffer similar restrictions).

No changes to the src\ambient_space\z2_visuals.jl script where maid, it needs update!

The general structure of the proyect is slightly changed, a __main__ module is still due.